refactor : contest.ts -> competition.ts 병합
