+++ front/src/pages/competitiondetail/participatewindow/ParticipateWindow.tsx
@@ -4,8 +4,6 @@ import { useParams, Link, Navigate, useNavigate } from 'react-router-dom'
 import { getTeamsinLeader } from '@/services/member'
 import Dropdown from '@/components/Dropdown/Dropdown'
 import Button from '@/components/Button/Button'
-import { CreateCompetitionDto } from '@/types/api'
-import { participateContest } from '@/services/contest'
 import { ParticipateTeamRequest } from '@/types/api'
 import logoIcon from '@svgs/logo/challSv2.svg'
 import ResultModal from './ResultModal/ResultModal'
