+++ front/src/pages/createcompetition/promotion/CompetitionModal/CompetitionModal.tsx
@@ -5,7 +5,7 @@ import { categorys } from '@/types/category'
 import Modal from '@components/Modal/Modal'
 import logoIcon from '@svgs/logo/challSv2.svg'
 import { CreateCompetitionDto, CompetitionCreateRequestDto } from '@/types/api'
-import { registerCompetition } from '@/services/contest'
+import { registerCompetition } from '@/services/competition'
 import { useResetRecoilState } from 'recoil'
 import { competitionForm, formButtonState } from '../../store'
 import ResultModal from '../ResultModal/ResultModal'
