+++ front/src/pages/modifycompetition/ModifyPromotion.tsx
@@ -2,7 +2,7 @@ import { useState, useRef, useEffect } from 'react'
 import { useParams } from 'react-router-dom'
 import { useQuery, useMutation, useQueryClient } from 'react-query'
 import Navbar from '@/components/Navbar/Navbar'
-import { getCompetitionDetails, modifyContestContent } from '@/services/contest'
+import { getCompetitionDetails, modifyContestContent } from '@/services/competition'
 import Button from '@/components/Button/Button'
 import CustomEditor from '@/components/Editor/CustomEditor'
 import { ApiResponse, UpdateCompetitionRequestDto } from '@/types/api'
