--- front/src/styles/index.tsx
@@ -1,2 +1,0 @@
-// 파일 구조를 위한 파일용이었음.
-// 작업할 때 필요하면 쓰고 아니면 삭제해도 됌
